+++ examples/static-sass/Cargo.toml
@@ -6,5 +6,4 @@ authors = ["Rasmus Kaj <kaj@kth.se>"]
 build = "src/build.rs"
 
 [build-dependencies]
-ructe = { path = "../.." }
-rsass = "*"
+ructe = { path = "../..", features = ["sass"] }
